Refactor commands to use a common CSV parser
